Rename uiNegotiator to uiApplyPatch
